requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Flask-Login==0.6.3
python-dotenv==1.0.0
werkzeug==3.0.0
gunicorn==21.2.0
